+++ src/Event/EventDispatcherConsumerInterface.php
@@ -1,0 +1,18 @@
+<?php
+namespace WoohooLabs\ApiFramework\Event;
+
+interface EventDispatcherConsumerInterface
+{
+    /**
+     * @param string $eventName
+     * @param string $className
+     * @param string $methodName
+     */
+    public function addClassListener($eventName, $className, $methodName);
+
+    /**
+     * @param string $eventName
+     * @param callable $listener
+     */
+    public function addCallbackListener($eventName, callable $listener);
+}