+++ src/Event/EventInterface.php
@@ -1,0 +1,15 @@
+<?php
+namespace WoohooLabs\ApiFramework\Event;
+
+interface EventInterface
+{
+    /**
+     * @return \WoohooLabs\ApiFramework\Request\RequestInterface
+     */
+    public function getRequest();
+
+    /**
+     * @return \WoohooLabs\ApiFramework\Response\ResponseInterface
+     */
+    public function getResponse();
+}