+++ src/EventConfig.php
@@ -1,0 +1,26 @@
+<?php
+namespace WoohooLabs\ApiFramework;
+
+class EventConfig extends Config
+{
+    /**
+     * @var \Closure|null
+     */
+    protected $events;
+
+    /**
+     * @return \Closure|null
+     */
+    public function getEvents()
+    {
+        return $this->events;
+    }
+
+    /**
+     * @param \Closure $events
+     */
+    public function setEvents(\Closure $events)
+    {
+        $this->events = $events;
+    }
+}